@@ -5,11 +5,13 @@ import (
 	"net/http"
 	"os"
 	"path/filepath"
-	"strings"
 
 	"github.com/gin-gonic/gin"
+	"github.com/open-cmi/cmmns/essential/events"
+	"github.com/open-cmi/cmmns/essential/i18n"
+	"github.com/open-cmi/cmmns/module/auditlog"
 	"github.com/open-cmi/cmmns/module/license"
-	"github.com/open-cmi/cmmns/pkg/path"
+	"github.com/open-cmi/cmmns/pkg/eyas"
 	"github.com/open-cmi/cmmns/service/webserver"
 )
 
@@ -30,36 +32,37 @@ func GetLicenseInfo(c *gin.Context) {
 }
 
 func UploadLicenseFile(c *gin.Context) {
+	ah := auditlog.NewAuditHandler(c)
 	file, _ := c.FormFile("file")
 	src, err := file.Open()
 	if err != nil {
+		ah.InsertOperationLog(i18n.Sprintf("upload license file"), false)
 		c.JSON(http.StatusOK, gin.H{"ret": -1, "msg": err.Error()})
 		return
 	}
 	defer src.Close()
 	content, err := io.ReadAll(src)
 	if err != nil {
+
+		ah.InsertOperationLog(i18n.Sprintf("upload license file"), false)
 		c.JSON(http.StatusOK, gin.H{"ret": -1, "msg": err.Error()})
 		return
 	}
-	arr := strings.Split(string(content), "\n")
-	if len(arr) < 2 {
-		c.JSON(http.StatusOK, gin.H{"ret": -1, "msg": "license format is invalid"})
-		return
-	}
-
-	licBase64 := arr[0]
-	signed := arr[1]
-	err = license.Verify(licBase64, signed)
+	os.Remove(file.Filename)
+	err = license.VerifyLicenseContent(string(content), true)
 	if err != nil {
-		c.JSON(http.StatusOK, gin.H{"ret": -1, "msg": "license signed string verified failed"})
+
+		ah.InsertOperationLog(i18n.Sprintf("upload license file"), false)
+		c.JSON(http.StatusOK, gin.H{"ret": -1, "msg": err.Error()})
 		return
 	}
 
-	rootdir := path.GetRootPath()
-	dst := filepath.Join(rootdir, "etc", "xsnos.lic")
+	confDir := eyas.GetConfDir()
+	dst := filepath.Join(confDir, "xsnos.lic")
 	out, err := os.OpenFile(dst, os.O_CREATE|os.O_RDWR, 0600)
 	if err != nil {
+
+		ah.InsertOperationLog(i18n.Sprintf("upload license file"), false)
 		c.JSON(http.StatusOK, gin.H{"ret": -1, "msg": err.Error()})
 		return
 	}
@@ -67,11 +70,14 @@ func UploadLicenseFile(c *gin.Context) {
 
 	_, err = out.Write(content)
 	if err != nil {
+
+		ah.InsertOperationLog(i18n.Sprintf("upload license file"), false)
 		c.JSON(http.StatusOK, gin.H{"ret": -1, "msg": err.Error()})
 		return
 	}
 
-	os.Remove(file.Filename)
+	ah.InsertOperationLog(i18n.Sprintf("upload license file"), true)
+	events.Notify("check-license-valid", nil)
 	c.JSON(http.StatusOK, gin.H{
 		"ret": 0,
 		"msg": "",
