@@ -1,5 +1,11 @@
 package subcommands
 
+import (
+	"fmt"
+	"os"
+	"path"
+)
+
 type HelpCommand struct {
 }
 
@@ -8,7 +14,13 @@ func (c *HelpCommand) Synopsis() string {
 }
 
 func (c *HelpCommand) Run() error {
-	Usage()
+	prog := path.Base(os.Args[0])
+	fmt.Printf("Usage: %s <subcommand> <subcommand args>\n\n", prog)
+
+	fmt.Printf("Subcommands:\n")
+	for name, command := range cmdmapping {
+		fmt.Printf("\t%-15s\t%s\n", name, command.Synopsis())
+	}
 	return nil
 }
 
