@@ -30,21 +30,10 @@ func RegisterCommand(key string, cmd Command) error {
 	return nil
 }
 
-func Usage() {
-	prog := path.Base(os.Args[0])
-	fmt.Printf("Usage: %s <subcommand> <subcommand args>\n\n", prog)
-
-	fmt.Printf("Subcommands:\n")
-	for name, command := range cmdmapping {
-		fmt.Printf("\t%-15s\t%s\n", name, command.Synopsis())
-	}
-	return
-}
-
 func Run() error {
+	prog := path.Base(os.Args[0])
 	if len(os.Args) < 2 {
-		Usage()
-		return errors.New("subcommand cli args is too short")
+		return fmt.Errorf("see '%s help'", prog)
 	}
 
 	var err error
@@ -53,7 +42,7 @@ func Run() error {
 	if ok {
 		err = cmd.Run()
 	} else {
-		err = RootCommand.Run()
+		return fmt.Errorf("'%s' is not a sub command\nsee '%s help'", os.Args[1], prog)
 	}
 
 	return err
