@@ -1,3 +1,3 @@
 package translation
 
-//go:generate gotext -srclang=en-US update -out=catalog.go -lang=en-US,zh-CN github.com/open-cmi/cmmns/main
+//go:generate gotext -srclang=en-US update -out=catalog-gen.go -lang=en-US,zh-CN github.com/open-cmi/cmmns/main
