@@ -1,11 +1,18 @@
 package main
 
 import (
+	"fmt"
+	"os"
+
 	"github.com/open-cmi/cmmns"
 
 	_ "github.com/open-cmi/cmmns/internal/translation"
 )
 
 func main() {
-	cmmns.Main()
+	err := cmmns.Main()
+	if err != nil {
+		fmt.Fprintln(os.Stderr, err)
+		os.Exit(1)
+	}
 }
