@@ -26,7 +26,7 @@ type Model struct {
 }
 
 func (m *Model) Save() error {
-	db := sqldb.GetConfDB()
+	db := sqldb.GetDB()
 
 	if m.isNew {
 		// 存储到数据库
@@ -65,7 +65,7 @@ func (m *Model) Save() error {
 }
 
 func (m *Model) Remove() error {
-	db := sqldb.GetConfDB()
+	db := sqldb.GetDB()
 
 	deleteClause := "delete from license where id=$1"
 	_, err := db.Exec(deleteClause, m.ID)
@@ -86,7 +86,7 @@ func New() *Model {
 
 func Get(id string) *Model {
 	queryClause := `select * from license where id=$1`
-	db := sqldb.GetConfDB()
+	db := sqldb.GetDB()
 	row := db.QueryRowx(queryClause, id)
 
 	var mdl Model
