@@ -11,8 +11,8 @@ import (
 )
 
 // List list
-func TokenList(option *goparam.Option) (int, []TokenRecord, error) {
-	db := sqldb.GetConfDB()
+func TokenList(option *goparam.Param) (int, []TokenRecord, error) {
+	db := sqldb.GetDB()
 
 	var results []TokenRecord = []TokenRecord{}
 
