@@ -9,7 +9,7 @@ import (
 )
 
 func RoleNameList() (int, []string, error) {
-	db := sqldb.GetConfDB()
+	db := sqldb.GetDB()
 
 	var roles []string = []string{}
 	countClause := "select count(*) from roles"
@@ -43,8 +43,8 @@ func RoleNameList() (int, []string, error) {
 	return count, roles, err
 }
 
-func RoleList(option *goparam.Option) (int, []Role, error) {
-	db := sqldb.GetConfDB()
+func RoleList(option *goparam.Param) (int, []Role, error) {
+	db := sqldb.GetDB()
 
 	var roles []Role = []Role{}
 	countClause := "select count(*) from roles"
@@ -83,7 +83,7 @@ func RoleList(option *goparam.Option) (int, []Role, error) {
 	return count, roles, err
 }
 
-func DeleteRole(option *goparam.Option, id string) error {
+func DeleteRole(option *goparam.Param, id string) error {
 	role := Get(id)
 	if role == nil {
 		return errors.New("role not exist")
