@@ -4,13 +4,18 @@ import (
 	"encoding/json"
 
 	"github.com/open-cmi/cmmns/essential/logger"
+	"github.com/open-cmi/cmmns/module/license"
 	"github.com/open-cmi/cmmns/module/system/prod"
 )
 
 func GetNavMenu(roleName string) []prod.Menu {
+	// 验证license
+	if !license.LicenseIsValid() {
+		return prod.GetRequireNav()
+	}
 	// 普通用户获取授权菜单
 	if roleName == "admin" {
-		menu := prod.GetProdNav()
+		menu := prod.GetNav()
 		return menu
 	}
 	var menu []prod.Menu = []prod.Menu{}
@@ -20,7 +25,7 @@ func GetNavMenu(roleName string) []prod.Menu {
 		return menu
 	}
 	if role.Permisions == "*" {
-		return prod.GetProdNav()
+		return prod.GetNav()
 	}
 	err := json.Unmarshal([]byte(role.Permisions), &menu)
 	if err != nil {
