@@ -12,7 +12,7 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/open-cmi/cmmns/essential/logger"
 	"github.com/open-cmi/cmmns/module/middleware"
-	"github.com/open-cmi/cmmns/pkg/path"
+	"github.com/open-cmi/cmmns/pkg/eyas"
 )
 
 type Service struct {
@@ -29,9 +29,9 @@ func (s *Service) Init() error {
 	// init webserver
 
 	middleware.DefaultMiddleware(s.Engine)
-	middleware.WACMiddleware(s.Engine)
+	//middleware.WACMiddleware(s.Engine)
 
-	workDir := path.GetRootPath()
+	workDir := eyas.GetRootPath()
 	dir := fmt.Sprintf("%s/static/", workDir)
 	s.Engine.Static("/api-static/", dir)
 	middleware.SessionMiddleware(s.Engine)
@@ -64,7 +64,7 @@ func (s *Service) Run() error {
 		} else if srv.Proto == "http" {
 			go s.Engine.Run(srv.Address + ":" + strconv.Itoa(srv.Port))
 		} else if srv.Proto == "https" {
-			rp := path.GetRootPath()
+			rp := eyas.GetRootPath()
 			var certFile string
 			if strings.HasPrefix(srv.CertFile, ".") {
 				certFile = filepath.Join(rp, srv.CertFile)
